parse_list_with_numpy helps to handle the restoration with the raw columns that has np.array inside them.
load_and_clean_results - upload one result.csv file and get it as a list of dicts, where each dict was a row in results.csv.

load_and_combine_results - use the previous, can take base_dir and search from there for all the results.csv and then insert them all to the same list of dicts: return a single, combined list of dictionaries. Each dictionary in the final list would represent a row from any of the CSV files.

load_all_results - similar to the previous but return list of list of dicts and dict of list of dicts.